Finished reworking ring plate task flow

+++ src/main/scala/edu/knuca/resmat/Api.scala
@@ -34,10 +34,12 @@ object Api extends App with Config with LazyLogging {
   val bucket = cfg.getString("bucket")
   val s3Manager = new S3Manager(accessKey, secretKey, bucket, "https://s3.eu-central-1.amazonaws.com")
 
-  if(MySql.migrateOnStartup) {
+  if(!Resmat.env.isProd && MySql.migrateOnStartup) {
     val flywayService = new FlywayService(MySql.flywayJdbcUrl, MySql.user, MySql.password, MySql.db, Flyway.baselineVersion)
-    flywayService.dropDatabase()
-    flywayService.migrateDatabaseSchema
+    if(!Resmat.env.isProd && MySql.dropSchemaOnStartup) {
+      flywayService.dropDatabase()
+    }
+    flywayService.migrateDatabaseSchema()
   }
 
   val mySqlDbConfig = DBConfig(MySql.jdbcUrl, MySql.user, MySql.password, MySql.driver, MySql.conns)
